+++ bee-block/tests/receipt_milestone_option.rs
@@ -185,7 +185,7 @@ fn pack_unpack_valid() {
     assert_eq!(packed_receipt.len(), receipt.packed_len());
     assert_eq!(
         receipt,
-        PackableExt::unpack_verified(&mut packed_receipt.as_slice()).unwrap()
+        PackableExt::unpack_verified(&mut packed_receipt.as_slice(), &()).unwrap()
     );
 }