+++ bee-ledger/bee-ledger-types/src/snapshot/milestone_diff.rs
@@ -52,6 +52,7 @@ impl MilestoneDiff {
 
 impl Packable for MilestoneDiff {
     type UnpackError = Error;
+    type UnpackVisitor = ();
 
     fn pack<P: Packer>(&self, packer: &mut P) -> Result<(), P::Error> {
         (self.milestone.packed_len() as u32 + MilestonePayload::KIND).pack(packer)?;
@@ -89,9 +90,10 @@ impl Packable for MilestoneDiff {
 
     fn unpack<U: Unpacker, const VERIFY: bool>(
         unpacker: &mut U,
+        visitor: &Self::UnpackVisitor,
     ) -> Result<Self, UnpackError<Self::UnpackError, U::Error>> {
-        let milestone_len = u32::unpack::<_, VERIFY>(unpacker).coerce()? as usize;
-        let payload = Payload::unpack::<_, VERIFY>(unpacker).coerce()?;
+        let milestone_len = u32::unpack::<_, VERIFY>(unpacker, visitor).coerce()? as usize;
+        let payload = Payload::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
         let milestone = match payload {
             Payload::Milestone(milestone) => milestone,
             _ => {
@@ -107,8 +109,8 @@ impl Packable for MilestoneDiff {
         }
 
         let consumed_treasury = if milestone.essence().options().receipt().is_some() {
-            let milestone_id = MilestoneId::unpack::<_, VERIFY>(unpacker).coerce()?;
-            let amount = u64::unpack::<_, VERIFY>(unpacker).coerce()?;
+            let milestone_id = MilestoneId::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
+            let amount = u64::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
 
             Some((
                 TreasuryOutput::new(amount).map_err(UnpackError::from_packable)?,
@@ -118,23 +120,23 @@ impl Packable for MilestoneDiff {
             None
         };
 
-        let created_count = u64::unpack::<_, VERIFY>(unpacker).coerce()?;
+        let created_count = u64::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
         let mut created_outputs = HashMap::with_capacity(created_count as usize);
 
         for _ in 0..created_count {
-            let output_id = OutputId::unpack::<_, VERIFY>(unpacker).coerce()?;
-            let created_output = CreatedOutput::unpack::<_, VERIFY>(unpacker).coerce()?;
+            let output_id = OutputId::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
+            let created_output = CreatedOutput::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
 
             created_outputs.insert(output_id, created_output);
         }
 
-        let consumed_count = u64::unpack::<_, VERIFY>(unpacker).coerce()?;
+        let consumed_count = u64::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
         let mut consumed_outputs = HashMap::with_capacity(consumed_count as usize);
 
         for _ in 0..consumed_count {
-            let output_id = OutputId::unpack::<_, VERIFY>(unpacker).coerce()?;
-            let created_output = CreatedOutput::unpack::<_, VERIFY>(unpacker).coerce()?;
-            let target = TransactionId::unpack::<_, VERIFY>(unpacker).coerce()?;
+            let output_id = OutputId::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
+            let created_output = CreatedOutput::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
+            let target = TransactionId::unpack::<_, VERIFY>(unpacker, visitor).coerce()?;
 
             consumed_outputs.insert(
                 output_id,