+++ bee-protocol/bee-protocol/CHANGELOG.md
@@ -33,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Updated dependencies;
+- Updated dependencies (including `packable`);
 
 ## 0.2.2 - 2022-03-07